[fidlc] Implement string_array type behind a flag
This is an implementation of the string_array type that is needed to
express kernel interfaces. It stores UTF-8 strings in-line and
zero-padded. This is gated behind the zx_c_types experimental flag,
though maybe it deserves its own flag.

It supports the syntax: string_array<N> to declare a string array type
and is represented in the JSON IR as a having kind==string_array and an
element_count reflecting the number of bytes.

Internal to fidlc it's represented as a constraint on array so that all
of the array layout logic can just flow naturally.

Change-Id: I331d981b082c4436f7a116671d6bdbe93250d3de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816062
Reviewed-by: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>