[netstack3] Get default BufferSizes from bindings
When a new TCP socket is created, derive the default send and receive
buffers from the NonSyncCtx, which is implemented in the bindings. This
fixes a bug introduced in I4088eab8067d8843a665fe7a2b29df6ce65ab716.

Before that change, the core provided no guidance on how to size the
receive buffer, so the bindings was using the zircon socket buffer size.
When that change introduced SO_RCVBUF support, it also added a default
receive buffer size in core. The new default size was smaller than the
value previously used by core, which caused a performance regression.

This CL moves the choice of default buffer sizes (both send and receive)
into the bindings, which allows the bindings to provide the same
original values (before the regression) to core when a new socket is
created.

Fixed: 123076
Change-Id: If3887d9590f293cdef2f5c1c04856be21754c498
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816268
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>