[fdf][registers] Migrate registers protocol phase1
This is phase 1 of the banjo to FIDL migration for the
fuchsia.hardware.registers protocol. This is a trampoline protocol and
there is an existing FIDL protocol by the same name. Since there are a
decent number of clients (some out of the main repo) this will be a soft
migration. That means we first have to support both.

This phase will:
 - Create a service under the FIDL for this protocol
 - Migrate the bind library's name to match the FIDL library
 - Support serving the FIDL service in the driver alongisde banjo

Tests: Client behavior is currently unchanged
Bug: 123422
Change-Id: I76062a1a366323c08ab29bce69508ec6c180de5c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/817329
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>