[fdf][ddk] Remove load_firmware_async
This API was added to solve a deadlock issue in a usb driver, but was
seemingly not used to solve that problem. It's seen no adoption outside
of tests since then. There seems to be no reason to keep it around.

Change-Id: Ie31ba019586ff885788a9b5996d9bf09e1880adf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/816727
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>