[async_patterns] PassDispatcher marker object to pass the dispatcher
It turns out some objects are not easily modified to obtain its
associated dispatcher via the TLS async_get_default_dispatcher. This CL
introduces a special PassDispatcher object that when supplied to
DispatcherBound as an argument to be sent, will turn into the correct
async_dispatcher_t* when it shows up on the other side at the async
object. This makes it easy to uplift an existing async object into the
DispatcherBound without changing it to require async-default.

Change-Id: I23dfbedb526c86851d91e49c2c7e29f7f68f113a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/817554
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>