[test_realm_proxy] Add a default test realm proxy service.
Change (3/3)

Adds a fuchsia_test_realm_proxy GN template that can be
used to define a canned implementation of a test realm proxy
service, providing access to a copies of some test realm component.
This is meant to serve as a basic implementation for the simplest
use cases.

Bug: 121395
Change-Id: I37b5b2d0f10778fd0b412d147b22603bfa37b658
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813748
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Kendal Harland <kjharland@google.com>