[tracing] Add producer targeting to perfetto-bridge
This change adds the ability to target an enabled category at
a specific perfetto producer to perfetto-bridge. It's modeled after
the ffx trace feature so category names take the form of
<producer_name>/<category>. This means that categories without a
producer prefix are applied to all producers, and categories with
a producer prefix are only applied to the producer matching that
prefix.

Change-Id: I45186100e53e4c579ce9295353e10ac73f7510c0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793129
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Bryant Chandler <bryantchandler@google.com>