[network] Fix GN arg to enable NS2 tracing
Now that the selection of which netstack to use is a part of product
assembly, the GN arg to enable tracing which switches the netstack
package in `//src/connectivity/network:netstack` no longer works for
enabling tracing in NS2.

Fix this by having the GN arg switch out the component in the
`netstack2` package (which is selected by product assembly) for the
component with tracing enabled when `enable_netstack2_tracing` is true.

Change-Id: Ibe8cafb9c0fefc4a545e80da17288a0c474386af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818110
Reviewed-by: Alex Konradi <akonradi@google.com>
Fuchsia-Auto-Submit: Tony Gong <gongt@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>