[media][codec] Add support for LC3 codec.
Most of the parameters required for LC3 encoder is in PcmFormat struct.
For the parameters that can't be represented by DomainFormat, we add
them to Lc3EncoderSettings. For decoder, the compressed audio table is
empty since most of the information can be embedded as part of
`oob_bytes`, similar to other codecs.

Bug: 122847
Change-Id: Ic10968ec4f50d3721467e4c11654b1c1f0963e47
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/812570
Reviewed-by: Dustin Green <dustingreen@google.com>
Commit-Queue: Dayeong Lee <dayeonglee@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
API-Review: Marie Janssen <jamuraa@google.com>