[assembly][recovery-like] Transition core_realm and core_realm_networking packages to assembly
Split the 'core_realm' AIB into two pieces, with the
AIB containing all '/core' children that are expected to be present when
the '/core' realm is present, and networking is present.

Change all packages in those two AIBs to be provided by Assembly, and
not the transitional labels that are being used to add them to the
recovery-like images.

Bug: 123397
Bug: 122862
Change-Id: Id37e9ab04274bbded1ed810d21b812e72c38c568
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/817230
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>