[audio] ADR: integrate driver_transfer_bytes
Remove any assumptions about linkage between internal_delay
and the required ring-buffer space a driver/device requires.
Internal_delay is still mandatory but is allowed to be zero.

Add driver_transfer_bytes to the logging and validation of
RingBufferProperties, and include it as the consumer_bytes
or producer_bytes, in the RingBuffer struct that we return.

When requesting the ring buffer VMO, only include the size
requested by the client -- do not also include the driver's
requirements. Also remove additional "gap" padding that was
previously included between consumer and producer sections.

Strengthen unittests for RingBufferProperties and DelayInfo.

Bug: 117887

Change-Id: Ia91577fc92df132ca319126b48be34f3067508e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818710
Reviewed-by: Tom Bergan <tombergan@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>