[netstack3] Plumb route metrics through ns3 core
This CL introduces the concept of a route metric to NS3 core. The metric
is set when creating a route, and can be later read back when listing
routes, or observing routing events.

Note that this CL does not introduce any changes to how routing
decisions are made. Honoring the route metric when sorting the routing
table will be introduced in a future change.

Bug: 122489
Change-Id: Ia16c5b3fe304d01cb3d64247896a3241bf8f3376
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/810928
Reviewed-by: Alex Konradi <akonradi@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Fuchsia-Auto-Submit: Jeff Martin <martinjeffrey@google.com>