[superproject] Roll third_party/pigweed pw_stream: SocketServer: allow reusing the port
Previously if the server started listening on port 0 it would not use
SO_REUSEADDR, so if later on the server was restarted on the same
(resolved) port number it would fail to bind. In order to fix we are
moving setsocktopt() call so that it happens for port 0 as well.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/133991
Original-Revision: c2df1e4942303b0e91f753f2e68b944b96fa9a93
GitOrigin-RevId: 22b9528bd2c5b37fab6cfcb4ba01ace3fcda9cd9
Roller-URL: https://ci.chromium.org/b/8786540421932659649
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I40bd999b5b12c9166d3f6cb1f224d6dc2f58aeb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819830