[kernel][vm] Check for no interval in merge and splice
Add checks for no intervals in the VmPageList MergeOnto/MergeFrom and
TakePages functions. To properly support intervals, we will need to
split partial intervals that are being merged to/from or spliced.
Currently the use cases for merging and splicing do not support
intervals in their page lists anyway, so keep the logic simple for now
and check that no intervals are present.

Bug: 122842
Run-All-Tests: true
Change-Id: Idcf582c04ce0e318ccf08350b8c4de0a9a50a633
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818733
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>