[starnix] Release the file descriptor of a Task at the time it dies
Currently, the fds are released at the time the Task id dropped, but
there is some case where the Task survives (because it is in a Arc)
while it is already dead. This ensures the fds can be released at the
right time.

This should be unnecessary when http://fxb/122600 is resolved.

Bug: 122600
Change-Id: I6240285b90d9d4993f1178dbb709aadbf3bb8551
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819807
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>