[ffx] Generate test data for ffx and subtools
This makes it so that when we build a subtool, we add a target that
generates 'test data' for it and puts it somewhere we can get at it.
This allows tests to not reproduce this effort every time they want to
call ffx tools.

Change-Id: I16650fd09478a7ab902e65183604d91d35007880
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/812172
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: David Turner <digit@google.com>