[magma][msd-vsi-vip] Fix WriteUnorderedEventIds
Reading and clearing the interrupt failed if the NNA was suspended.

The test 'WriteUnorderedEventIds' bypasses SubmitCommandBuffer. This can
result in the device suspending in the middle of the test.

Added an option to DeviceThread to disable suspend.
This option is for use in testing.

Change logic to for hangcheck timeout to ensure suspend timeout cannot
trigger a hang.

Change debug messages in interrupt routine.

tested: msd-vsi-vip-integration-test
bug: fxbug.dev/123433
Change-Id: I8254a80cf39d5a876f8edf5e0499da22a5d08cb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819189
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: Jamie Howarth <jhowarth@google.com>