[fshost-rust] Reset fvm partitions on format
This CL ensures that after the account migrations are complete and
we switch over to hardware backed keys, the FVM is reset to a known
good state such that the account and virtualization partitions are
destroyed and the FVM data partition is resized accordingly to take
up the newly freed space on disk. Note that as part of this, the data
partition is destroyed and reallocated so data loss is incurred.

A bug was also uncovered during this CL which is worth documenting.
When fvm_ramdisk is set, there is a ZxcryptMatcher that will attach the
zxcrypt driver to the data device and this new ZxcryptDevice will be
picked up by the device handler and run through various match_device()
implementations associated with different matchers. There is a moment
after we have destroyed the data device in reset_fvm_partition where
the PartitionMapMatcher will call content_format() on this ZxcryptDevice
and because the device has already been destroyed, the get_info() call
which eventually calls `out_info->block_count -= info_.reserved_blocks;`
will cause block_count to underflow and later cause a multiplication
overflow in content_format(). To avoid this, we now clamp block_count
to 0 if it underflows.

Fixed: 119223
Change-Id: Icc7a2791ab57d1e9e33887f3b2182cd1dee19f17
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/791309
Fuchsia-Auto-Submit: Nikita Jindal <nikitajindal@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>