[fake-display] Make fake-display not a system driver anymore.
fake-display should be only allowed to use within a fake
device tree provided by fake-hardware-display-controller-provider
in a test realm (as specified in Fuchsia RFC-0180: Test UI Stack).
It shouldn't be used as a system driver in a non-hermetic environment.

This removes the driver binding logic, the driver component and the
fake display "device" specification so that fake-display will not be
able to bind to any device.

Test: core.qemu-arm64 builds
Fixed: 123816
Change-Id: I3fa272d4a278eecca2bee39a3d6eeac2d9a78ac4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/820834
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>