[benchmarks] Split tracing benchmarks out from microbenchmarks_test
One reason for doing this is to allow flakes in these sets of tests to
be tracked separately (by flake-fetcher and other tools), because we
have found a couple of flaky failure modes that affect the tracing
tests recently.

This also has the benefit of making these tests a little more
convenient to run locally:

 * tracing_microbenchmarks_test is much faster to run on its own.

 * microbenchmarks_test doesn't depend on the SL4F server now
   (following the recent change to runTestComponent()).

Note that we just moved microbenchmarks_test.dart to
tracing_microbenchmarks_test.dart and removed a small part, but
Git/Gerrit do not recognise that move.

Bug: 123479
Test: fx set terminal.x64 --release
  + fx test --e2e host_x64/tracing_microbenchmarks_test
Change-Id: I066ad7debf100313a3b0cfbd7ac1e260a17a5dad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/819210
Commit-Queue: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>