[netstack] Enable inline static init optimization
This was disabled in http://go.dev/cl/467015 due to bugs that cause
compilation failures. As we are not affected by these bugs, it's safe to
turn this on and avoid the 300KiB bloat that is otherwise caused by
upgrading Go from 1.20 to 1.20.2.

Change-Id: I291b39e3572964c99700199d95ca05d8f3214914
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821088
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>