[ctf] Add Subpackaged rust_test_runner Shard
Adds a shard to include the rust-test-runner and lib_loader_runner_cache
(a dependency of all test runners) as subpackages. This shard defines an
environment where the subpackaged test runner exists. Tests using this
runner should start the test component in this environment. This will
enable Rust tests in the CTF by ensuring that the Rust test runner is
compatible with the test.

Bug: 122914

Change-Id: I8e1c7f500b330ef571375eaf9ff168307b48bd06
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/807612
Commit-Queue: John Shamoon <johnshamoon@google.com>
Reviewed-by: Kendal Harland <kjharland@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>