[display][core] Replace DoublyLinkedListNode with fbl NodeState.
This replaces the previous DoublyLinkedListNode which contains
a RefPtr to the Image itself, with fbl NodeState class, which
encapsulates all the RefPtr manipulation, making it safer for
developers to use.

Note that since the node has to be mutex-guarded, we cannot
directly use DoublyLinkedListable traits, instead we have to
add a mutex-guarded DoublyLinkedListNode member to Image. The
implementation is the same, but the methods are checked by
compile-time thread-safety checker.

Test: display-core-unittests
Fixed: 123301
Change-Id: Id0e9b6525061dd7b518e7efb9e255729c890adef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818351
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>