[starnix] Do not swallow binder errors
Binder errors are currently, in the vast majority, an issue with Starnix
code. Because of this, and to make progress, they were currently
swallowed, and a success was returned to the client code.

The starnix binder code is now in a better situation than previously,
and the user mode code has been modified to crash loudly in case of
binder error.

It is now beneficial to see these crashes to be able to correlate client
code and binder code, so the error swallowing mechanism has been
removed.

Change-Id: I50477da7e5fdd5b8d8db6fdfd53c9e950ed63d11
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821789
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>