[xhci] Fix issues with large size transfers
Transfer descriptor size can be more than 31 for large data transfers,
but needs to be set as 31 for anything above it. The physical buffers
can be more than 64KiB in which case it  needs to be split into different
TRBs as the max length per TRB is 64KiB. This CL fixes both of these
issues.

Fixed: 107831, 122933
Change-Id: I6c2a27736be1cf773e90df108be7105b618d6b1f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/814410
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>