[fshost] remove fshost_rust feature checks
When deleting the cpp version of these tests, the fshost_rust feature
was removed because it was no longer necessarily to differentiate
between the two. However, checks for that feature still existed and
would always come back false. This adds back the behavior for the rust
version unconditionally. Luckily this is just some additional checking
and nothing load-bearing.

Change-Id: I6cfbda2d63e3ab331499df57d597a51d00a6dffc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818870
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>