[starnix] Fix signal stack underflow
When setting up the signal stack, it was possible for the stack
computation to underflow, causing a panic in the rust code. The code now
does checked arithmetic and handles the error.

The error is handled by delivering a SIGSEGV signal to the application.
If SIGSEGV is masked, it will be unmasked; if it is ignored, the signal
disposition will be reset to default before delivering the signal. If
the application has a handler for SIGSEGV, the signal will be delivered
to the handler, and if it fails, the signal disposition will be reset to
default and the signal redelivered.

Test: fx test syscalls_test
Co-Authored-By: marinaciocea@google.com
Bug: 118658
Change-Id: I05d827af007c11f733c411bc95f6254931508421
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802168
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Commit-Queue: Marina Ciocea <marinaciocea@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>