[fxfs] Unwrap keys in the background
Unwrapping keys in the background when opening a file significantly
reduces the time it takes to respond to open calls. The keys are only
necessary for encrypting and decrypting the contents of a file. The keys
will almost always be ready before the contents of the file need to be
encrypted or decrypted.

Fixed: 123394
Change-Id: I6a72018c1fb11bc6d3b81b63082ca3efdcc56705
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/818589
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>