[trace_manager] Add a Timeout to List Categories
Unlike Start and Stop, list categories waits for a response from the
trace provider. If a trace provider does not respond, this will cause
trace_manager to wait forever for it.

This patch races each GetKnownCategories response against a timeout, and
defaults the response to and empty list if the trace provider does not
respond it time.

Bug: 123170
Change-Id: Ie041b0e38d05096948293f40ce2fb8fe3550b7b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821172
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Brian Hamrick <bhamrick@google.com>