[core-tests] Allow slightly longer timeout
As we attempt to move to `fastboot boot` in infra, we discovered that
the core tests run slower on some devices when booted via `fastboot
boot` instead of being mexec'ed from zedboot.  This is because firmware
on said devices does not place the cores in the maximum performance
state, whereas once zedboot has booted, the `power_manager` there does,
and core-tests (by virtue of relying ~only on the kernel) inherit the
processor power state of whatever loaded them.  In particular, we
observed on sherlock ASAN builds, the core-tests took 634 seconds when
launched from fastboot, which exceeds the 600 second default timeout.

Options included:

* change firmware to use the maximum p-state always (risky, due to
  potentially changing boot ordering, which would be painful to
  re-qualify)
* change firmware to use the maximum p-state only when using
  `fastboot boot` (less risky but causes behavior divergence)
* chainload through some other boot shim (plausible but probably more
  work than makes sense)
* teach core-tests to crank up to the maximum p-state
  (awkward: setting p-states is complicated and platform-specific, and
  even if it's safe now, it may not be safe for all hardware to sit
  indefinitely at max p-state)
* just accept the slightly worse performance on core-tests, and increase
  the timeout

The last option seemed the most supportable at this point in time, and
it unblocks further rollout of ffx-in-infra.

Fixed: 124069
Change-Id: I2783e1f1af867e2bf41ea64772739f0d9f0a45d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/822206
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>