[starnix] Release the file descriptor of a Task at the time it dies
Currently, the fds are released at the time the Task id dropped, but
there is some case where the Task survives (because it is in a Arc)
while it is already dead. This ensures the fds can be released at the
right time.

This should be unnecessary when http://fxb/122600 is resolved.

This is a reland of https://fuchsia-review.googlesource.com/c/fuchsia/+/819807
The changes are:
- The file descriptors are releases after the task is removed from the
  ThreadGroup to prevent races due to closed files.
- Use a named variable to force the ordering of dropped.

Bug: 122600
Change-Id: I335e43f147efee0d2e50d7b3c186d4dcb747127b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/821888
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>