[kernel][x86][boot-shim] Add a IKCONFIG to the linuxboot boot-shim
Some emulators use extract-ikconfig to determine the architecture of the
boot-shim, therefore we must insert a very minimal IKCONFIG into the
boot-shim.

extract-ikconfig greps for IKCFG_ST to find the config in the kernel.

ikconfig.txt is the contents of the config, and it is gzipped to produce
ikconfig.gz. Because this file is so small we decided to check it in
instead of adding a gzip tool.

Change-Id: I951ba2b56f4f63f4cab49188beae307b51582908
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/812288
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>